added i, d and u specifier functions and tests for mandatory and bonus part. Part of bonus flags are still not correctly working

+++ ft_printf/src/ft_printf_char.c
@@ -6,7 +6,7 @@
 /*   By: sguilher <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/11 23:09:22 by sguilher          #+#    #+#             */
-/*   Updated: 2021/08/17 23:02:27 by sguilher         ###   ########.fr       */
+/*   Updated: 2021/08/18 05:27:51 by sguilher         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,9 +39,9 @@ void	printf_c(t_print *p, t_flags *f, char c)
 
 void	printf_s(t_print *p, t_flags *f, char *s)
 {
-	size_t	len;
+	int	len;
 
-	len = ft_strlen(s);
+	len = (int)ft_strlen(s);
 	while((*f).width > len)
 	{
 		printf_putchar_fd(p, ' ');
